app(2.2.eng.ver)/backend/requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0

# Database
psycopg2==2.9.11

# Authentication
djangorestframework-simplejwt==5.3.0

# Environment Variables
python-decouple==3.8
python-dotenv==1.0.0

# Image Processing
Pillow==12.0.0

# API Client
requests==2.31.0
groq==0.4.1

# Utilities
pytz==2023.3
python-dateutil==2.8.2

# Development
django-extensions==3.2.3
ipython==8.17.2

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Code Quality
black==23.11.0
flake8==6.1.0
isort==5.12.0
